+++ src/runner.sh
@@ -142,6 +142,11 @@ function runner::run_test() {
 
   if [[ "$current_assertions_failed" != "$(state::get_assertions_failed)" ]]; then
     state::add_tests_failed
+
+    if [ "$STOP_ON_FAILURE" = true ]; then
+      exit 1
+    fi
+
     return
   fi